# Make NodeList readonly struct and optimize indexers