--- a/samples/Esprima.Benchmark/Properties/AssemblyInfo.cs
+++ b/samples/Esprima.Benchmark/Properties/AssemblyInfo.cs
@@ -1,5 +1,4 @@
 ﻿using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 // General Information about an assembly is controlled through the following