--- a/src/Esprima/Ast/ArrayPattern.cs
+++ b/src/Esprima/Ast/ArrayPattern.cs
@@ -4,15 +4,17 @@ namespace Esprima.Ast
 {
     public class ArrayPattern : Node, BindingPattern
     {
-        public readonly NodeList<IArrayPatternElement> Elements;
+        private readonly NodeList<IArrayPatternElement> _elements;
 
-        public ArrayPattern(NodeList<IArrayPatternElement> elements) :
+        public ArrayPattern(in NodeList<IArrayPatternElement> elements) :
             base(Nodes.ArrayPattern)
         {
-            Elements = elements;
+            _elements = elements;
         }
 
+        public ref readonly NodeList<IArrayPatternElement> Elements => ref _elements;
+
         public override IEnumerable<INode> ChildNodes =>
-            ChildNodeYielder.Yield(Elements);
+            ChildNodeYielder.Yield(_elements);
     }
 }