--- a/src/Esprima/Ast/ArrowParameterPlaceHolder.cs
+++ b/src/Esprima/Ast/ArrowParameterPlaceHolder.cs
@@ -5,16 +5,17 @@ namespace Esprima.Ast
     public class ArrowParameterPlaceHolder : Node, Expression
     {
         public static readonly ArrowParameterPlaceHolder Empty = new ArrowParameterPlaceHolder(new NodeList<INode>());
+        private readonly NodeList<INode> _params;
 
-        public readonly NodeList<INode> Params;
-
-        public ArrowParameterPlaceHolder(NodeList<INode> parameters) :
+        public ArrowParameterPlaceHolder(in NodeList<INode> parameters) :
             base(Nodes.ArrowParameterPlaceHolder)
         {
-            Params = parameters;
+            _params = parameters;
         }
 
+        public ref readonly NodeList<INode> Params => ref _params;
+
         public override IEnumerable<INode> ChildNodes =>
-            ChildNodeYielder.Yield(Params);
+            ChildNodeYielder.Yield(_params);
     }
 }