--- a/src/Esprima/Ast/CallExpression.cs
+++ b/src/Esprima/Ast/CallExpression.cs
@@ -5,21 +5,22 @@ namespace Esprima.Ast
     public class CallExpression : Node,
         Expression
     {
+        private readonly NodeList<ArgumentListElement> _arguments;
         public readonly Expression Callee;
-        public readonly NodeList<ArgumentListElement> Arguments;
-
         public bool Cached;
         public bool CanBeCached = true;
         public object CachedArguments;
 
-        public CallExpression(Expression callee, NodeList<ArgumentListElement> args) :
+        public CallExpression(Expression callee, in NodeList<ArgumentListElement> args) :
             base(Nodes.CallExpression)
         {
             Callee = callee;
-            Arguments = args;
+            _arguments = args;
         }
 
+        public ref readonly NodeList<ArgumentListElement> Arguments => ref _arguments;
+
         public override IEnumerable<INode> ChildNodes =>
-            ChildNodeYielder.Yield(Callee, Arguments);
+            ChildNodeYielder.Yield(Callee, _arguments);
     }
 }