--- a/src/Esprima/Ast/CallExpression.cs
+++ b/src/Esprima/Ast/CallExpression.cs
@@ -6,6 +6,7 @@ namespace Esprima.Ast
         Expression
     {
         private readonly NodeList<ArgumentListElement> _arguments;
+
         public readonly Expression Callee;
         public bool Cached;
         public bool CanBeCached = true;