--- a/src/Esprima/Ast/ExportNamedDeclaration.cs
+++ b/src/Esprima/Ast/ExportNamedDeclaration.cs
@@ -4,19 +4,24 @@ namespace Esprima.Ast
 {
     public class ExportNamedDeclaration : Node, ExportDeclaration
     {
+        private readonly NodeList<ExportSpecifier> _specifiers;
         public readonly IStatementListItem Declaration;
-        public readonly NodeList<ExportSpecifier> Specifiers;
         public readonly Literal Source;
 
-        public ExportNamedDeclaration(IStatementListItem declaration, NodeList<ExportSpecifier> specifiers, Literal source) :
+        public ExportNamedDeclaration(
+            IStatementListItem declaration,
+            in NodeList<ExportSpecifier> specifiers,
+            Literal source) :
             base(Nodes.ExportNamedDeclaration)
         {
             Declaration = declaration;
-            Specifiers = specifiers;
+            _specifiers = specifiers;
             Source = source;
         }
 
+        public ref readonly NodeList<ExportSpecifier> Specifiers => ref _specifiers;
+
         public override IEnumerable<INode> ChildNodes =>
-            ChildNodeYielder.Yield((INode) Declaration, Specifiers, Source);
+            ChildNodeYielder.Yield(Declaration, _specifiers, Source);
     }
 }