--- a/src/Esprima/Ast/ExportNamedDeclaration.cs
+++ b/src/Esprima/Ast/ExportNamedDeclaration.cs
@@ -5,6 +5,7 @@ namespace Esprima.Ast
     public class ExportNamedDeclaration : Node, ExportDeclaration
     {
         private readonly NodeList<ExportSpecifier> _specifiers;
+
         public readonly IStatementListItem Declaration;
         public readonly Literal Source;