--- a/src/Esprima/Ast/NodeList.cs
+++ b/src/Esprima/Ast/NodeList.cs
@@ -38,7 +38,6 @@ namespace Esprima.Ast
         }
 
         public NodeList<INode> AsNodes() =>
-            // Conversion by co-variance!
             new NodeList<INode>(_items /* conversion by co-variance! */, _count);
 
         public T this[int index]