--- a/src/Esprima/Ast/SequenceExpression.cs
+++ b/src/Esprima/Ast/SequenceExpression.cs
@@ -4,13 +4,15 @@ namespace Esprima.Ast
 {
     public class SequenceExpression : Node, Expression
     {
-        public NodeList<Expression> Expressions { get; internal set; }
+        internal NodeList<Expression> _expressions;
 
-        public SequenceExpression(NodeList<Expression> expressions) :
+        public SequenceExpression(in NodeList<Expression> expressions) :
             base(Nodes.SequenceExpression)
         {
-            Expressions = expressions;
+            _expressions = expressions;
         }
+
+        public ref readonly NodeList<Expression> Expressions => ref _expressions;
 
         public override IEnumerable<INode> ChildNodes =>
             ChildNodeYielder.Yield(Expressions);