--- a/src/Esprima/Ast/ThisExpression.cs
+++ b/src/Esprima/Ast/ThisExpression.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using System.Linq;
 
 namespace Esprima.Ast
 {