… src/Esprima/Ast/TemplateLiteral.cs
… src/Esprima/Ast/VariableDeclaration.cs
… src/Esprima/ExceptionHelper.cs
… src/Esprima/HoistingScope.cs
… src/Esprima/JavascriptParser.cs
… src/Esprima/JitHelper.cs
… src/Esprima/Utils/AstJson.cs