As `NodeList` is passed around and used a lot on Jint's side, it makes sense to change from regular `struct` to `readonly struct`. Now there will be less defensive copies when methods (like the indexer) are being used. I also checked the indexers and made them use half the IL code to make them perform better.

Here are benchmark results from Jint where the Esprma usage is fraction of the time used running, but we can see positive boost nevertheless.

## Jint.Benchmark.DromaeoBenchmark

| **Diff**|Method|FileName|Mean|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|-------:|
| Old |Run|dromaeo-3d-cube|59.53 ms|1222.2222|222.2222|-|7417.87 KB|
| **New** |	|	| **59.20 ms (-1%)** | **1222.2222 (0%)** | **222.2222 (0%)** | **-** | **7417.87 KB (0%)** |
| Old |Run|dromaeo-core-eval|13.13 ms|78.1250|-|-|377.66 KB|
| **New** |	|	| **13.04 ms (-1%)** | **78.1250 (0%)** | **-** | **-** | **377.66 KB (0%)** |
| Old |Run|dromaeo-object-array|154.16 ms|41500.0000|1500.0000|250.0000|174933.85 KB|
| **New** |	|	| **151.00 ms (-2%)** | **41500.0000 (0%)** | **1500.0000 (0%)** | **250.0000 (0%)** | **174932.15 KB (0%)** |
| Old |Run|droma(...)egexp [21]|717.10 ms|50000.0000|30000.0000|20000.0000|407356.26 KB|
| **New** |	|	| **697.65 ms (-3%)** | **53000.0000 (+6%)** | **33000.0000 (+10%)** | **21000.0000 (+5%)** | **408566.76 KB (0%)** |
| Old |Run|droma(...)tring [21]|769.28 ms|90000.0000|49000.0000|46000.0000|1378209.27 KB|
| **New** |	|	| **771.58 ms (0%)** | **108000.0000 (+20%)** | **67000.0000 (+37%)** | **64000.0000 (+39%)** | **1378167.91 KB (0%)** |


## Jint.Benchmark.StopwatchBenchmark

| **Diff**|Method|N|Mean|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|-------:|
| Old |Jint|1|913.1 ms|12000.0000|-|-|51.61 MB|
| **New** |	|	| **894.3 ms (-2%)** | **12000.0000 (0%)** | **-** | **-** | **51.61 MB (0%)** |