@lahma 2019-07-23
Thank you @atifaziz for the review, good observations. I've incorporated your feedback and it should now follow your recommendations, except I've kept the ExceptionHelper and didn't split another PR for the indexing optimization (I'm lazy yes, but I did create a separate commit though 😉 ).