@atifaziz 2019-07-24
> I've incorporated your feedback and it should now follow your recommendations

Tremendous and thanks!

> …except I've kept the ExceptionHelper

[Too bad](https://github.com/sebastienros/esprima-dotnet/pull/109#discussion_r307018321) I couldn't convince you otherwise but okay.

> …and didn't split another PR for the indexing optimization (I'm lazy yes, but I did create a separate commit though 😉 ).

Thumbs-up for a separate commit but it made the review longer nonetheless. This was not a complaint but feedback for the future to help speed up reviews. When one is looking at pages of diffs, it's difficult to be lazy as a reviewer because you're building a mental stack only to unstack and re-stack, if you see what I mean. I don't know how @sebastienros does it, frankly. I need to sit next to him once and learn.

Thanks again for all the good and hard work in this PR!